+++ CHANGELOG.md
@@ -1,9 +1,19 @@
 # Change Log
 
-All notable changes to the "visualizer-validation-rules-salesforce" extension will be documented in this file.
+All notable changes to the "salesforce-validation-rules-viewer" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
 
 - Initial release
+
+## [1.0.2] - 2025-05-27
+
+### Added
+
+- Emojis in commands
+
+### Changed
+
+- Extension Name Change